Restrict permissions by using XDG desktop portals
Since 7.3.0, it is possible to use XDG desktop portals to open files.
This means the "host" permissions for filesystems is not needed anymore.

Also removes the permissions for /tmp as unlike what is reported in #16,
it is not needed anymore to ensure a single instance of OnlyOffice is
running.

Restrict the permission to devices to "dri" (for hardware acceleration).
Dunno why this was set to "all" in the first place.